requirements.txt

discord.py==2.3.2
python-dotenv==1.0.1
aiohttp==3.9.1
audioop-lts==0.2.1
psycopg2-binary
